installed: webpack, babel and configurations
