+++ index.html
@@ -75,7 +75,7 @@
     <script src='app/PopupImage.js'></script>
     <script src='app/FormValidator.js'></script>
     <script src='app/UserInfo.js'></script>
-    <script src='app/script.js'></script>
+    <script src='app/index.js'></script>
 
 </body>
