+++ package.json
@@ -1,0 +1,33 @@
+{
+    "name": "yp-webpack",
+    "version": "1.0.0",
+    "description": "Yandex Praktikum Project 11",
+    "main": "index.js",
+    "scripts": {
+        "build": "webpack --mode production",
+        "dev": "webpack-dev-server --mode development --open --watch"
+    },
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/BigBaraBum/yp-webpack.git"
+    },
+    "author": "Nikita Inkin",
+    "license": "ISC",
+    "bugs": {
+        "url": "https://github.com/BigBaraBum/yp-webpack/issues"
+    },
+    "homepage": "https://github.com/BigBaraBum/yp-webpack#readme",
+    "devDependencies": {
+        "@babel/cli": "^7.8.4",
+        "@babel/core": "^7.8.4",
+        "@babel/preset-env": "^7.8.4",
+        "babel-loader": "^8.0.6",
+        "webpack": "^4.41.6",
+        "webpack-cli": "^3.3.11",
+        "webpack-dev-server": "^3.10.3"
+    },
+    "dependencies": {
+        "babel-polyfill": "^6.26.0",
+        "core-js": "^3.4.1"
+    }
+}
