+++ webpack.config.js
@@ -1,0 +1,15 @@
+const path = require('path');
+module.exports = {
+    entry: { main: './app/index.js' },
+    output: {
+        path: path.resolve(__dirname, 'dist'),
+        filename: 'main.js'
+    },
+    module: {
+        rules: [{ // тут описываются правила
+            test: /\.js$/, // регулярное выражение, которое ищет все js файлы
+            use: { loader: "babel-loader" }, // весь JS обрабатывается пакетом babel-loader
+            exclude: /node_modules/ // исключает папку node_modules
+        }]
+    }
+}
